+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-project(bot2-core)
+project(robotlocomotion-lcmtypes)
 cmake_minimum_required(VERSION 2.8.12)
 
 # TODO remove when minimum CMake >= 3.7
