--- pkgconfig/lcmtypes_bot2-core.pc.in
@@ -1,12 +1,0 @@
-prefix=@BOT2_CORE_INSTALL_PREFIX@
-exec_prefix=${prefix}
-libdir=${exec_prefix}/lib@LIB_SUFFIX@
-includedir=${prefix}/include
-
-Name: lcmtypes_bot2-core
-Description: LCM types for bot2-core
-Requires:  lcm
-Version: 0.0.0
-Libs: -L${libdir} -llcmtypes_bot2-core
-Cflags: -I${includedir}
-classpath=
