+++ pkgconfig/robotlocomotion-lcmtypes.pc.in
@@ -1,0 +1,12 @@
+prefix=@ROBOTLOCOMOTION-LCMTYPES_INSTALL_PREFIX@
+exec_prefix=${prefix}
+libdir=${exec_prefix}/lib@LIB_SUFFIX@
+includedir=${prefix}/include
+
+Name: robotlocomotion-lcmtypes
+Description: LCM types for robotlocomotion
+Requires:  lcm
+Version: 0.0.0
+Libs: -L${libdir} -lrobotlocomotion-lcmtypes
+Cflags: -I${includedir}
+classpath=
